Use consistent integer types for view positions and sizes
Actual window sizes and positions fit easily in a 16-bit integer.  So, we use
that in "representation contexts" like events.  This makes structures smaller,
and allows the values to be converted to float, double, or integer without
casting (since any int16_t or uint16_t value can fit in them without loss).

Setter APIs use native integers for convenience, to avoid casting hassles when
doing arithmetic.  Ranges are checked at runtime.

+++ src/mac_cairo.m
@@ -89,7 +89,10 @@ puglMacCairoEnter(PuglView* view, const PuglExposeEvent* expose)
   const double scale = 1.0 / [[NSScreen mainScreen] backingScaleFactor];
   CGContextRef context =
     (CGContextRef)[[NSGraphicsContext currentContext] graphicsPort];
-  const CGSize sizePx = {view->frame.width, view->frame.height};
+
+  const CGSize sizePx = {(CGFloat)view->frame.width,
+                         (CGFloat)view->frame.height};
+
   const CGSize sizePt = CGContextConvertSizeToUserSpace(context, sizePx);
 
   // Convert coordinates to standard Cairo space
